pyproject.toml
[tool.poetry]
name = "chapa-bot"
version = "0.1.0"
description = ""
authors = ["hunderaweke <hunderaweke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "13.14"
python-dotenv = "^1.0.0"
fastapi = "^0.103.1"
pydantic = "^2.4.1"
pyperclip3 = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
